Extract InterpreterRuntime
